+++ src/server.rs
@@ -110,7 +110,6 @@ impl Plugin for ServerPlugin {
             .add_observer(handle_connects)
             .add_observer(handle_disconnects)
             .add_observer(buffer_despawns)
-            .add_systems(Startup, setup_channels)
             .add_systems(
                 PreUpdate,
                 (
@@ -175,10 +174,14 @@ impl Plugin for ServerPlugin {
             app.register_required_components::<ConnectedClient, ReplicatedClient>();
         }
     }
-}
 
-fn setup_channels(mut server: ResMut<RepliconServer>, channels: Res<RepliconChannels>) {
-    server.setup_client_channels(channels.client_channels().len());
+    fn finish(&self, app: &mut App) {
+        app.world_mut()
+            .resource_scope(|world, mut server: Mut<RepliconServer>| {
+                let channels = world.resource::<RepliconChannels>();
+                server.setup_client_channels(channels.client_channels().len());
+            });
+    }
 }
 
 /// Increments current server tick which causes the server to replicate this frame.