+++ tests/mutations.rs
@@ -33,7 +33,8 @@ fn small_component() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -79,7 +80,8 @@ fn package_size_component() {
                 ..Default::default()
             }),
         ))
-        .replicate::<VecComponent>();
+        .replicate::<VecComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -132,7 +134,8 @@ fn many_components() {
             }),
         ))
         .replicate::<BoolComponent>()
-        .replicate::<VecComponent>();
+        .replicate::<VecComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -186,7 +189,8 @@ fn once() {
                 ..Default::default()
             }),
         ))
-        .replicate_once::<BoolComponent>();
+        .replicate_once::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -233,7 +237,8 @@ fn periodic() {
                 ..Default::default()
             }),
         ))
-        .replicate_periodic::<BoolComponent>(PERIOD);
+        .replicate_periodic::<BoolComponent>(PERIOD)
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -289,7 +294,8 @@ fn periodic_with_miss() {
                 ..Default::default()
             }),
         ))
-        .replicate_periodic::<BoolComponent>(PERIOD);
+        .replicate_periodic::<BoolComponent>(PERIOD)
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -353,7 +359,8 @@ fn related() {
             }),
         ))
         .sync_related_entities::<ChildOf>()
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -398,7 +405,8 @@ fn command_fns() {
             }),
         ))
         .replicate::<OriginalComponent>()
-        .set_command_fns(replace, command_fns::default_remove::<ReplacedComponent>);
+        .set_command_fns(replace, command_fns::default_remove::<ReplacedComponent>)
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -450,7 +458,8 @@ fn marker() {
         .set_marker_fns::<ReplaceMarker, _>(
             replace,
             command_fns::default_remove::<ReplacedComponent>,
-        );
+        )
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -512,7 +521,8 @@ fn marker_with_history() {
             write_history,
             command_fns::default_remove::<BoolComponent>,
         )
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -588,7 +598,8 @@ fn marker_with_history_consume() {
             command_fns::default_remove::<BoolComponent>,
         )
         .replicate::<BoolComponent>()
-        .replicate::<MappedComponent>();
+        .replicate::<MappedComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -676,7 +687,8 @@ fn marker_with_history_old_update() {
             write_history,
             command_fns::default_remove::<BoolComponent>,
         )
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -744,7 +756,8 @@ fn many_entities() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -797,7 +810,8 @@ fn with_insertion() {
             }),
         ))
         .replicate::<BoolComponent>()
-        .replicate::<TestComponent>();
+        .replicate::<TestComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -841,7 +855,8 @@ fn with_removal() {
             }),
         ))
         .replicate::<BoolComponent>()
-        .replicate::<TestComponent>();
+        .replicate::<TestComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -884,7 +899,8 @@ fn with_despawn() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -932,7 +948,8 @@ fn buffering() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -989,7 +1006,8 @@ fn old_ignored() {
                 ..Default::default()
             }),
         ))
-        .replicate::<MappedComponent>();
+        .replicate::<MappedComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -1056,7 +1074,8 @@ fn acknowledgment() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -1132,9 +1151,9 @@ fn confirm_history() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
-    client_app.finish();
 
     server_app.connect_client(&mut client_app);
 
@@ -1198,7 +1217,8 @@ fn after_disconnect() {
                 ..Default::default()
             }),
         ))
-        .replicate::<BoolComponent>();
+        .replicate::<BoolComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);